pyproject.toml
[tool.poetry]
name = "datagears"
version = "0.1.0"
description = "Large-scale data processing and inference engine for machine learning."
authors = ["sam <contact@justsam.io>"]
license = "BSD 3-Clause License"

[tool.poetry.dependencies]
python = "^3.9"
networkx = "^2.5"
matplotlib = "^3.3.4"
pydot = "^1.4.2"
redis = "^3.5.3"
cloudpickle = "^1.6.0"
sanic-jsonrpc = "^0.3.1"
grpcio = "^1.37.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
black = "^20.8b1"
isort = "^5.8.0"
grpcio-tools = "^1.37.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
